Moved updateLastLocation from ManagerMetadataUtil to AbstractTabletStateStore
